refactor: server.js into handlers and transmit
